+++ x/ccv/difftest/relay.go
@@ -1,4 +1,4 @@
-package pbt
+package difftest
 
 import (
 	"bytes"
